feat: add /form route, .env file, change to ES Modules
