--- .vercel/project.json
@@ -1,1 +1,0 @@
-{"orgId":"team_sRi9f5M7302HBJ4QFNDp8cJc","projectId":"prj_dW3qv8XbwcpQ9NJMca6hNP9XCVbp"}
